working on trial

+++ packages/runtime/src/sim/actions/save_reload.rs
@@ -79,18 +79,7 @@ pub fn reload(
 ) -> Result<(), processor::Error> {
     // can only reload from inventory (System tab)
     super::switch_to_inventory_or_stop!(ctx, sys, errors, "RELOAD");
-    regen_stage(ctx, sys, errors, Some(load_gdt), true)
-}
-
-
-
-/// Simulate a loading screen, such as entering/exiting shrine
-pub fn loading_screen(
-    ctx: &mut sim::Context<&mut Cpu2>,
-    sys: &mut sim::GameSystems,
-    errors: &mut Vec<ErrorReport>,
-) -> Result<(), processor::Error> {
-    regen_stage(ctx, sys, errors, None, false)
+    regen_stage_internal(ctx, sys, errors, true, Some(load_gdt))
 }
 
 
@@ -105,17 +94,19 @@ pub fn reload_gdt(
 }
 
 /// Regenerate the game stage, and optionally load a save while doing that
-pub fn regen_stage(
+pub fn regen_stage_internal(
     ctx: &mut sim::Context<&mut Cpu2>,
     sys: &mut sim::GameSystems,
     errors: &mut Vec<ErrorReport>,
+    remove_translucent: bool,
     load_save: Option<&gdt::TriggerParam>,
-    load_gdt: bool
 ) -> Result<(), processor::Error> {
-    // 0. Translucent items are removed before reload
-    // This is called as part of the loading screen, for example,
-    // this is called when entering/exiting shrine as well
-    linker::delete_removed_items(ctx.cpu())?;
+    if remove_translucent {
+        // 0. Translucent items are removed before reload
+        // This is called as part of the loading screen, for example,
+        // this is called when entering/exiting shrine as well
+        linker::delete_removed_items(ctx.cpu())?;
+    }
 
     // 1. BaseProcMgr deletes all actors
     sys.overworld.destroy_all();
@@ -126,24 +117,28 @@ pub fn regen_stage(
     // 2. SaveMgr/GdtMgr (?) loads the save into GDT
     if let Some(save_gdt) = load_save {
         reload_gdt_or_stop!(ctx, errors, save_gdt);
-    }
-
-    // 3. PMDM loads from GDT
-    if load_gdt {
+        // 3. PMDM loads from GDT
         linker::load_from_game_data(ctx.cpu())?;
     }
 
     // 4. Create player equipments
+    recreate_overworld_equipments(ctx, sys)?;
+
+    Ok(())
+}
+
+pub fn recreate_overworld_equipments(
+    ctx: &mut sim::Context<&mut Cpu2>,
+    sys: &mut sim::GameSystems,
+) -> Result<(), processor::Error> {
     let state = linker::events::CreateEquip::execute_subscribed(
         ctx.cpu(),
         CreateEquipState::default(),
         CreateEquipState::update,
         linker::create_player_equipment,
     )?;
-
-    // 4. Equipments update their value
+    // update value to PMDM
     sys.overworld.reload_equipments(ctx.cpu(), state.weapon, state.bow, state.shield)?;
-
     Ok(())
 }
 
@@ -153,8 +148,6 @@ pub fn trial_start(
     sys: &mut sim::GameSystems,
     errors: &mut Vec<ErrorReport>,
 ) -> Result<(), processor::Error> {
-    // 0. To avoid issues, we switch to overworld
-    super::switch_to_overworld_or_stop!(ctx, sys, errors, "!TRIAL-START");
 
     // X. Delete overworld weapons
     // I think this is part of createPlayerEquipment, but we don't
@@ -175,8 +168,6 @@ pub fn trial_start(
     // 3. Equipments update their value
     sys.overworld.reload_equipments(ctx.cpu(), state.weapon, state.bow, state.shield)?;
 
-    // TODO: we don't know yet how refights work
-
     Ok(())
 }
 
@@ -187,6 +178,7 @@ pub fn trial_finish(
 ) -> Result<(), processor::Error> {
     // 0. To avoid issues, we switch to overworld
     super::switch_to_overworld_or_stop!(ctx, sys, errors, "!TRIAL-FINISH");
+    todo!()
 }
 
 #[derive(Default)]
